+++ src/gui/widgets/cpuload_widget.cpp
@@ -2,7 +2,7 @@
  * cpuload_widget.cpp - widget for displaying CPU-load (partly based on
  *                      Hydrogen's CPU-load-widget)
  *
- * Copyright (c) 2005-2007 Tobias Doerffel <tobydox/at/users.sourceforge.net>
+ * Copyright (c) 2005-2014 Tobias Doerffel <tobydox/at/users.sourceforge.net>
  * 
  * This file is part of Linux MultiMedia Studio - http://lmms.sourceforge.net
  *
@@ -90,7 +90,7 @@ void cpuloadWidget::paintEvent( QPaintEvent *  )
 void cpuloadWidget::updateCpuLoad()
 {
 	// smooth load-values a bit
-	Uint8 new_load = ( m_currentLoad + engine::mixer()->cpuLoad() ) / 2;
+	int new_load = ( m_currentLoad + engine::mixer()->cpuLoad() ) / 2;
 	if( new_load != m_currentLoad )
 	{
 		m_currentLoad = new_load;
